+++ src/com/company/repo/MatchRepoService.java
@@ -1,0 +1,9 @@
+package com.company.repo;
+
+import com.company.bean.MatchBean;
+import com.company.dto.MatchDto;
+
+public interface MatchRepoService {
+
+    void matchDetailUpdation(MatchBean matchBean) ;
+}
