+++ src/com/company/repo/PerBallDetailsRepoService.java
@@ -1,0 +1,11 @@
+package com.company.repo;
+
+import com.company.bean.PerBallBean;
+import com.company.dto.MatchDto;
+
+import java.util.ArrayList;
+
+public interface PerBallDetailsRepoService {
+
+    void perBallDetailUpdation(ArrayList<PerBallBean> perBallStatus) ;
+}
