+++ src/com/company/repo/TeamDetailsRepoService.java
@@ -1,0 +1,9 @@
+package com.company.repo;
+
+import com.company.bean.TeamBean;
+import com.company.dto.MatchDto;
+
+public interface TeamDetailsRepoService {
+
+    void teamDetailUpdation(TeamBean teamBean);
+}
