+++ src/com/company/services/ScoreBoardService.java
@@ -3,7 +3,7 @@ package com.company.services;
 import com.company.dto.MatchDto;
 
 public interface ScoreBoardService {
-    void showLiveScore(MatchDto thisMatchScoreBoard);
+    void showLiveScore(MatchDto thisMatchScoreBoard, int teamId);
     void showFinalResult(MatchDto thisMatchScoreBoard);
     void showTeam1ScoreCard(MatchDto thisMatchScoreBoard);
     void showTeam2ScoreCard(MatchDto thisMatchScoreBoard);
