+++ src/com/company/util/DbConnectorUtil.java
@@ -1,18 +1,19 @@
 package com.company.util;
 
+import com.company.variables.EnvironmentVariables;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
 
 public class DbConnectorUtil {
     public static Connection getConnection() {
         try {
-            String driver = "com.mysql.cj.jdbc.Driver";
-            String dataBaseUrl = "jdbc:mysql://localhost:3306/cricketData";
-            String username = "root";
-            String password = "password";
+            String driver = EnvironmentVariables.driver ;
+            String dataBaseUrl = EnvironmentVariables.dataBaseUrl;
+            String username = EnvironmentVariables.username;
+            String password = EnvironmentVariables.password;
             Class.forName(driver);
             Connection conn = DriverManager.getConnection(dataBaseUrl, username, password);
-//            System.out.println("Database connected sucessfully");
             return conn;
         }
         catch (Exception e) {
