+++ src/com/company/variables/EnvironmentVariables.java
@@ -1,0 +1,10 @@
+package com.company.variables;
+
+public class EnvironmentVariables {
+
+    public static String driver = "com.mysql.cj.jdbc.Driver";
+    public static String dataBaseUrl = "jdbc:mysql://localhost:3306/cricketData";
+    public static String username = "root";
+    public static String password = "password";
+
+}
